+++ config/application.rb
@@ -13,6 +13,7 @@ require "action_mailer/railtie"
 require "action_view/railtie"
 require "action_cable/engine"
 # require "rails/test_unit/railtie"
+require_relative '../lib/middleware/authenticate_token'
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -39,7 +40,7 @@ module PostbeeApi
     # Only loads a smaller set of middleware suitable for API only apps.
     # Middleware like session, flash, cookies can be added back manually.
     # Skip views, helpers and assets when generating a new resource.
-    config.middleware.use 'AuthenticateToken'
+    config.middleware.use AuthenticateToken
     config.api_only = true
   end
 end
